+++ src/App.js
@@ -12,14 +12,14 @@ import Signup from './components/authentication/signup';
 import ForgotPassword from './components/authentication/forgotPassword';
 import ResetPassword from './components/authentication/resetPassword';
 import SuccessPage from './components/authentication/successPage';
-import Discussion from './components/pages/Discussion';
-import PastShifts from './components/pages/PastShifts';
+import Discussion from './components/pages/discussion/Discussion';
+import PastShifts from './components/pages/pastShifts/PastShifts';
 import Contacts from './components/pages/contacts/Contacts';
-import History from './components/pages/History';
+import History from './components/pages/history/History';
 import PrivateRoute from './components/authentication/PrivateRoute';
-import ShiftSignUp from './components/pages/ShiftSignUp';
-import DiscussionThread from './components/pages/DiscussionThread';
-import ScheduleManager from './components/pages/ScheduleManager';
+import ShiftSignUp from './components/pages/schedule/ShiftSignUp';
+import DiscussionThread from './components/pages/discussion/DiscussionThread';
+import ScheduleManager from './components/pages/schedule/ScheduleManager';
 
 // Your web app's Firebase configuration
 // For Firebase JS SDK v7.20.0 and later, measurementId is optional
