+++ src/components/pages/discussion/DiscussionPost.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import styled from 'styled-components';
-import { GreyDiv } from '../../styled-components/discussion-components';
+import { GreyDiv } from '../../../styled-components/discussion-components';
 
 const Author = styled.span`
   font-size: 18px;
