+++ src/components/pages/pastShifts/PastShifts.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import HeaderWithNav from '../navigation/nav-header';
+import HeaderWithNav from '../../navigation/nav-header';
 
 export default function PastShifts(props) {
   const { isAdmin } = props;
